simulations/raw-data-generator/requirements.txt
paho-mqtt
sqlalchemy
psycopg2-binary
python-dotenv